Implemented do syntax.
