+++ src/main/java/de/michab/scream/Procedure.java
@@ -164,26 +164,26 @@ public class Procedure
         return apply( parent, evaluatedArgs );
     }
 
-    @Override
-    public Thunk _activate( Environment e, Cons args, Cont<FirstClassObject> c )
-            throws RuntimeX
-    {
-        checkArgumentCount( args );
-
-        final var ex = e.extend( getName() );
-
-        if ( _rest != Cons.NIL )
-            ex.define( (Symbol)_rest, Cons.NIL );
-
-        Cont<Cons> cc = (cons) -> _bind(
-                ex,
-                _formalArguments,
-                cons,
-                s -> Continuation._x_begin( s, _body, c ) );
-
-        return () -> Continuation._x_eval( e, args, cc );
-    }
-
+//    @Override
+//    public Thunk _activate_( Environment e, Cons args, Cont<FirstClassObject> c )
+//            throws RuntimeX
+//    {
+//        checkArgumentCount( args );
+//
+//        final var ex = e.extend( getName() );
+//
+//        if ( _rest != Cons.NIL )
+//            ex.define( (Symbol)_rest, Cons.NIL );
+//
+//        Cont<Cons> cc = (cons) -> _bind(
+//                ex,
+//                _formalArguments,
+//                cons,
+//                s -> Continuation._x_begin( s, _body, c ) );
+//
+//        return () -> Continuation._x_eval( e, args, cc );
+//    }
+//
     /**
      * This is the public available apply procedure that is used from Scheme
      * implemented methods via reflection.
