+++ src/main/java/de/michab/scream/pops/SyntaxDo.java
@@ -1,12 +1,16 @@
 package de.michab.scream.pops;
 
+import org.smack.util.JavaUtil;
+
 import de.michab.scream.Cons;
 import de.michab.scream.Environment;
 import de.michab.scream.FirstClassObject;
+import de.michab.scream.Lambda;
 import de.michab.scream.RuntimeX;
 import de.michab.scream.ScreamException.Code;
 import de.michab.scream.Symbol;
 import de.michab.scream.Syntax;
+import de.michab.scream.util.Scut;
 
 /**
  * <code>
@@ -52,6 +56,95 @@ public class SyntaxDo extends Syntax
     }
 
     /**
+     *
+     * @param variables {@code ((<var1> <init1> opt<step1>) ... )}
+     * @return Two lists: The car-list contains the init-expressions.  The cdr-list
+     * contains the step-expressions.
+     * @throws RuntimeX
+     */
+    private Cons validateVariables( Cons variables ) throws RuntimeX
+    {
+
+        Cons inits = Cons.NIL;
+        Cons steps = Cons.NIL;
+
+        while ( variables != Cons.NIL )
+        {
+            var variableDef =
+                    Scut.as( Cons.class, variables.getCar() );
+            checkArgumentCount( 2, 3, variableDef );
+
+            Symbol variable =
+                    Scut.as( Symbol.class, variableDef.listRef( 0 ) );
+
+            // Add a new init pair.
+            inits = new Cons(
+                    Cons.create(
+                            variable,
+                            variableDef.listRef( 1 )), inits );
+
+            if ( variableDef.length() == 3 )
+            {
+                // Add a new step pair.
+                steps = new Cons(
+                        Cons.create(
+                                variable,
+                                variableDef.listRef( 2 )), steps );
+            }
+
+            variables = (Cons)variables.getCdr();
+        }
+
+        return new Cons( inits, steps );
+    }
+
+    @Override
+    protected Lambda _compile( Environment env, Cons args ) throws RuntimeX
+    {
+        checkArgumentCount( 2, Integer.MAX_VALUE, args );
+
+        var variables =
+                Scut.as( Cons.class, args.listRef( 0 ) );
+        var test =
+                Scut.as( Cons.class, args.listRef( 1 ) );
+        // May be nil.
+        var commands =
+                args.listTail( 2 );
+
+        var setup =
+                validateVariables( variables );
+
+        checkArgumentCount( 1, Integer.MAX_VALUE, test );
+
+
+        return new Lambda(
+                (e,c) -> Continuation._x_do(
+                        e,
+                        (Cons)setup.getCar(),
+                        (Cons)setup.getCdr(),
+                        test,
+                        commands,
+                        c ),
+                this.getName() ).setInfo( args );
+    }
+
+    @Override
+    public FirstClassObject compile( Environment parent, Cons args )
+            throws RuntimeX
+    {
+        return _compile( parent, args );
+    }
+    @Override
+    public FirstClassObject activate( Environment parent,
+            Cons arguments )
+                    throws RuntimeX
+    {
+        var λ = _compile( parent, arguments );
+
+        return FirstClassObject.evaluate( λ, parent );
+    }
+
+    /**
      * Checks if the passed argument is a <code>Cons</code>, is not NIL and is
      * a proper list.  Transforms the list into an array and returns that.
      * @throws RuntimeX
@@ -69,6 +162,7 @@ public class SyntaxDo extends Syntax
     public FirstClassObject compile( Environment parent, FirstClassObject[] args )
             throws RuntimeX
     {
+        JavaUtil.Assert( false, "Unexpected call to SyntaxDo#compile()" );
         checkMinimumArgumentCount( 2, args );
 
         try
