+++ src/main/resources/de/michab/scream/Scream.properties
@@ -29,7 +29,6 @@ integrateClasses = \
   de.michab.scream.pops.SyntaxCond \
   de.michab.scream.pops.SyntaxDefine \
   de.michab.scream.pops.SyntaxDo \
-  de.michab.scream.pops.SyntaxDo2 \
   de.michab.scream.pops.SyntaxIf \
   de.michab.scream.pops.SyntaxLambda \
   de.michab.scream.pops.SyntaxLet \
